working save game editor

+++ forge-gui-mobile-dev/src/forge/app/EditSave.java
@@ -27,12 +27,17 @@ public class EditSave extends ApplicationAdapter {
     private final int saveId;
     private final byte clearFlagValue;
     private final byte travelFlagValue;
-    private final byte is
+    private final byte isDryRun;
+    private final String clearStatus;
+    private final String travelStatus;
 
     public static void main(String[] args) {
         int saveId = Integer.parseInt(args[0]);
         byte clearFlagValue = Byte.parseByte(args[1]);
         byte travelFlagValue = Byte.parseByte(args[2]);
+        byte isDryRun = Byte.parseByte(args[3]);
+        String clearStatus = args[4];
+        String travelStatus = args[5];
 
         // This should fix MAC-OS startup without the need for -XstartOnFirstThread parameter
         if (SharedLibraryLoader.isMac) {
@@ -43,7 +48,7 @@ public class EditSave extends ApplicationAdapter {
         Localizer.getInstance().initialize("en-US", assetsDir + "res/languages/");
         String switchOrientationFile = assetsDir + "switch_orientation.ini";
 
-        ApplicationListener start = Forge.getApp(new EditSave(saveId, clearFlagValue, travelFlagValue), new Lwjgl3Clipboard(), new Main.DesktopAdapter(switchOrientationFile),
+        ApplicationListener start = Forge.getApp(new EditSave(saveId, clearFlagValue, travelFlagValue, isDryRun, clearStatus, travelStatus), new Lwjgl3Clipboard(), new Main.DesktopAdapter(switchOrientationFile),
                 assetsDir, false, false, 0, false, 0, "", "");
 
         FModel.initialize(null, null);
@@ -52,10 +57,13 @@ public class EditSave extends ApplicationAdapter {
         new Lwjgl3Application(start, config);
     }
 
-    public EditSave(int saveId, byte clearFlagValue, byte travelFlagValue) {
+    public EditSave(int saveId, byte clearFlagValue, byte travelFlagValue, byte isDryRun, String clearStatus, String travelStatus) {
         this.saveId = saveId;
         this.clearFlagValue = clearFlagValue;
         this.travelFlagValue = travelFlagValue;
+        this.isDryRun = isDryRun;
+        this.clearStatus = clearStatus;
+        this.travelStatus = travelStatus;
     }
 
     @Override
@@ -88,9 +96,11 @@ public class EditSave extends ApplicationAdapter {
                 for (AdventureQuestStage stage : quest.stages) {
                     if (stage.name.equals("Clear")) {
                         stage.mapFlagValue = clearFlagValue;
+                        stage.setStatus(clearStatus);
                     }
                     else if (stage.name.equals("Travel")) {
                         stage.mapFlagValue = travelFlagValue;
+                        stage.setStatus(travelStatus);
                     }
                 }
                 break;
@@ -98,6 +108,12 @@ public class EditSave extends ApplicationAdapter {
         }
 
         System.out.println(player.toJson());
+
+        if (isDryRun != 0) {
+            System.out.println("Dry Run - exiting");
+            return;
+        }
+        System.out.println("Saving file");
 
         try {
             try(FileOutputStream fos =  new FileOutputStream(fileName);
@@ -109,6 +125,8 @@ public class EditSave extends ApplicationAdapter {
 
                 oos.writeObject(mainData);
             }
+
+            System.out.println("Success!");
 
         } catch (IOException e) {
             e.printStackTrace();
